refactor: divide contents in groupApiClient.ts (#171)
