+++ src/lib/api/groupApiClient.ts
@@ -1,4 +1,4 @@
-import { IContainer, IFood, IGroup } from '@/types/definition';
+import { IGroup } from '@/types/definition';
 
 import { request } from './commonUtils';
 
@@ -9,31 +9,9 @@ const BACKEND_API_DOMAIN = process.env.NEXT_PUBLIC_BACKEND_API_DOMAIN || '';
  * @property {string} groupId - The unique identifier of the group.
  * @property {string} groupName - The name of the group.
  */
-interface IGroupApiResponse {
+export interface IGroupApiResponse {
   groupId: string;
   groupName: string;
-}
-
-/**
- * Interface representing the API response for a container.
- * @property {string} id - The unique identifier of the container.
- * @property {string} name - The name of the container.
- * @property {IGroupApiResponse} group - The group to which this container belongs.
- * @property {IFood[]} foods - Array of foods contained in this container.
- */
-interface IContainerApiResponse {
-  id: string;
-  name: string;
-  group: IGroupApiResponse;
-  foods: IFood[];
-}
-
-/**
- * Interface representing the API response for multiple containers.
- * @property {IContainerApiResponse[]} containers - Array of container objects.
- */
-interface IContainersResponse {
-  containers: IContainerApiResponse[];
 }
 
 /**
@@ -58,54 +36,6 @@ export const fetchGroupList = async (): Promise<IGroup[]> => {
       id: group.groupId,
       name: group.groupName,
     }));
-  } catch (err) {
-    throw new Error('API response is invalid'); // TODO: display error page
-  }
-};
-
-/**
- * Converting a Container object from the Backend API to a front Interface
- * @param containers array of containers with Backend API schema
- * @returns array of containers with IContainer
- */
-const convertApiResponseContainers = (containers: IContainerApiResponse[]): IContainer[] => {
-  return containers.map((container: IContainerApiResponse) => ({
-    ...container,
-    group: {
-      id: container.group.groupId,
-      name: container.group.groupName,
-    },
-  }));
-};
-
-/**
- * Fetch an array of containers associated with a certain group .
- * @param id groupId
- * @returns Array of IContainer object
- */
-export const fetchContainerList = async (id: string): Promise<IContainer[]> => {
-  try {
-    const data: IContainersResponse = await request<IContainersResponse>({
-      url: BACKEND_API_DOMAIN + '/groups/' + id + '/containers',
-      method: 'GET',
-    });
-    return convertApiResponseContainers(data.containers);
-  } catch (err) {
-    throw new Error('API response is invalid'); // TODO: display error page
-  }
-};
-
-/**
- * Fetch an array of containers associated with a logged-in user.
- * @returns Array of IContainer object
- */
-export const fetchAllContainerList = async (): Promise<IContainer[]> => {
-  try {
-    const data: IContainersResponse = await request<IContainersResponse>({
-      url: BACKEND_API_DOMAIN + '/containers',
-      method: 'GET',
-    });
-    return convertApiResponseContainers(data.containers);
   } catch (err) {
     throw new Error('API response is invalid'); // TODO: display error page
   }
