+++ src/lib/api/index.ts
@@ -1,3 +1,4 @@
 export { request } from './commonUtils';
-export { fetchAllContainerList, fetchContainerList, fetchGroupList } from './groupApiClient';
+export { fetchAllContainerList, fetchContainerList } from './containerApiClient';
+export { fetchGroupList } from './groupApiClient';
 export { fetchUserList } from './userApiClient';
